requirements.txt
attrs
numpy
tblib
tritonclient[grpc]
